requirements.txt
aiogram==3.18.0
alembic==1.14.1
async-timeout==5.0.1
asyncpg==0.30.0
docx2txt==0.9
dropbox==12.0.2
email_validator==2.2.0
fastapi-cli==0.0.7
firebase-admin==6.6.0
google-auth-oauthlib==1.2.1
grpcio-status==1.70.0
httptools==0.6.4
Jinja2==3.1.5
jsonschema==4.23.0
langchain-chroma==0.2.3
langchain-experimental==0.3.4
langchain-google-genai==2.1.2
langchain-openai==0.3.13
langgraph==0.3.29
msal==1.31.1
pip==25.0.1
pypdf==5.4.0
pytest-mock==3.14.0
python-multipart==0.0.20
redis==5.2.1
sqlmodel==0.0.22
tonyg-rfc3339==0.1
watchfiles==1.0.4
websockets==14.2
